% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents to a map of a specific state for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number describing the state to plot}

\item{year}{A number describing the year to plot}
}
\value{
This function returns a map plot of a given state visualizing
   the accidents based on their location
}
\description{
The "fars_map_state" function reads yearly data of FARS datasets,
filters the data to the given state number, and plots the accidents based on 
their geographic locations (latitude, longitude). You can customize the state 
number and the years of which you want to plot (using the \code{state.num} and 
\code{year} arguments). The function uses the "make_filename" and the 
"fars_read" functions to read the datasets for the year specified, then filters
the data for the given state number, and plots the accidents to a map of 
the specific state based on the locations of accidents.
}
\note{
There is an "if" statement verifying that the given \code{state.num}
   is a valid state number. Stops if it's invalid. Additionally, 
   another "if" statement checks if the state-filtered dataset 
   contains any accidents. If it's empty, gives an error message and
   stops with an invisible NULL result.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
fars_map_state(8, 2015)
}
}

