% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read "Fatality Analysis Reporting System" datasets}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A file path where the dataset is available}
}
\value{
This function returns a dataset as a "data.frame" (and also incorporates
   the dplyr package's "tbl_df" and "tbl" classes). Additionally, it prints out
   information about the dataset to the console if a variable name to save is 
   not defined.
}
\description{
The "fars_read" function reads a csv format FARS dataset. You can customize
the file to read (using the \code{filename} argument). If you do not save
it to a variable, the function will print out information about the dataset
to the console.
}
\note{
There is an "if" condition in the beginning, which checks if the given 
   file path exists, results in an error if returns FALSE.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
fars.df <- fars_read("accident_2014.csv.bz2")
}
}

