% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read "Fatality Analysis Reporting System" data specifying the year number}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A number or vector of numbers describing the years to analyze}
}
\value{
This function returns a list containing the data tables with two 
   variables ("MONTH", "year").
}
\description{
The "fars_read_years" function reads yearly data of FARS datasets 
into a list where each element of the list contains specific data for the given year. 
You can customize the years of which you want to get more information (using
the \code{years} argument). The function reads the specified FARS datasets and 
mutates a new variable ("year") for the certain year, and selects the "MONTH"
attribute from the dataset, then gives a list of data tables of these two attributes
as a result.
}
\note{
There is an error handling with a tryCatch() function. In case of any
   errors, it results in a warning.
}
\examples{
\dontrun{
fars_read_years(2015)
fars_read_years(c(2013, 2014))
fars.year <- fars_read_years(2013)
}
}

