% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents by month and year in "Fatality Analysis 
Reporting System" data specifying the year numbers}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A number or vector of numbers describing the years to analyze}
}
\value{
This function returns a dataframe describing the amount of accidents
   for the given years.
}
\description{
The "fars_summarize_years" function reads yearly data of FARS datasets 
and summarizes the number of accidents by month for each year specified. You can 
customize the years of which you want to get more information (using the 
\code{years} argument). The function uses the "fars_read_years" function to read 
year-specific FARS data, then groups the observations by year and month, and 
summarizes the number of accidents into a dataframe.
}
\examples{
\dontrun{
fars_summarize_years(2015)
fars_summarize_years(c(2013, 2014))
fars.summary <- fars_summarize_years(2013)
}
}

