% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a default filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number or vector of numbers describing the years. It can also 
be given as character values.}
}
\value{
This function returns a character vector, which is in the default 
   format of the FARS filenames.
}
\description{
The "make_filename" function creates a character vector, which 
describes the default filename for FARS files. You can customize the year in 
the filename (using the \code{year} argument). It saves the year to an 
integer variable inside the function, then concatenates the given year
number with the "accident" sting and includes the file format, too.
}
\examples{
make_filename(2013)
make_filename("2013")
make_filename(c(2013, 2014))

}

